requirements.txt
django==4.2.9
psycopg[binary]
faker
django-debug-toolbar==4.2.0
